Convert HttpServletRequest to Sentry Request once for all events triggered during request execution.
