+++ sentry-spring/src/main/java/io/sentry/spring/SentryRequestHttpServletRequestProcessor.java
@@ -13,22 +13,17 @@ import org.jetbrains.annotations.Nullable;
 @Open
 public class SentryRequestHttpServletRequestProcessor implements EventProcessor {
   private final @NotNull HttpServletRequest request;
-  private final @NotNull SentryRequestResolver sentryRequestResolver;
   private final @NotNull TransactionNameProvider transactionNameProvider =
       new TransactionNameProvider();
 
   public SentryRequestHttpServletRequestProcessor(
-      final @NotNull HttpServletRequest request,
-      final @NotNull SentryRequestResolver sentryRequestResolver) {
+      final @NotNull HttpServletRequest request) {
     this.request = Objects.requireNonNull(request, "request is required");
-    this.sentryRequestResolver =
-        Objects.requireNonNull(sentryRequestResolver, "sentryRequestResolver are required");
   }
 
   @Override
   public @NotNull SentryEvent process(
       final @NotNull SentryEvent event, final @Nullable Object hint) {
-    event.setRequest(sentryRequestResolver.resolveSentryRequest(request));
     if (event.getTransaction() == null) {
       event.setTransaction(transactionNameProvider.provideTransactionName(request));
     }
