+++ sentry-spring/src/test/kotlin/io/sentry/spring/SentryRequestHttpServletRequestProcessorTest.kt
@@ -25,116 +25,11 @@ class SentryRequestHttpServletRequestProcessorTest {
 
         fun getSut(request: HttpServletRequest, options: SentryOptions = SentryOptions()): SentryRequestHttpServletRequestProcessor {
             whenever(hub.options).thenReturn(options)
-            return SentryRequestHttpServletRequestProcessor(request, SentryRequestResolver(hub))
+            return SentryRequestHttpServletRequestProcessor(request)
         }
     }
 
     private val fixture = Fixture()
-
-    @Test
-    fun `attaches basic information from HTTP request to SentryEvent`() {
-        val request = MockMvcRequestBuilders
-            .get(URI.create("http://example.com?param1=xyz"))
-            .header("some-header", "some-header value")
-            .accept(MediaType.APPLICATION_JSON)
-            .buildRequest(MockServletContext())
-        val eventProcessor = fixture.getSut(request)
-        val event = SentryEvent()
-
-        eventProcessor.process(event, null)
-
-        assertNotNull(event.request) {
-            assertEquals("GET", it.method)
-            assertEquals(mapOf(
-                "some-header" to "some-header value",
-                "Accept" to "application/json"
-            ), it.headers)
-            assertEquals("http://example.com", it.url)
-            assertEquals("param1=xyz", it.queryString)
-        }
-    }
-
-    @Test
-    fun `attaches header with multiple values`() {
-        val request = MockMvcRequestBuilders
-            .get(URI.create("http://example.com?param1=xyz"))
-            .header("another-header", "another value")
-            .header("another-header", "another value2")
-            .buildRequest(MockServletContext())
-        val eventProcessor = fixture.getSut(request)
-        val event = SentryEvent()
-
-        eventProcessor.process(event, null)
-
-        assertNotNull(event.request) {
-            assertEquals(mapOf(
-                "another-header" to "another value,another value2"
-            ), it.headers)
-        }
-    }
-
-    @Test
-    fun `when sendDefaultPii is set to true, attaches cookies information`() {
-        val request = MockMvcRequestBuilders
-            .get(URI.create("http://example.com?param1=xyz"))
-            .header("Cookie", "name=value")
-            .header("Cookie", "name2=value2")
-            .buildRequest(MockServletContext())
-        val sentryOptions = SentryOptions()
-        sentryOptions.isSendDefaultPii = true
-        val eventProcessor = fixture.getSut(request, sentryOptions)
-        val event = SentryEvent()
-
-        eventProcessor.process(event, null)
-
-        assertNotNull(event.request) {
-            assertEquals("name=value,name2=value2", it.cookies)
-        }
-    }
-
-    @Test
-    fun `when sendDefaultPii is set to false, does not attach cookies`() {
-        val request = MockMvcRequestBuilders
-            .get(URI.create("http://example.com?param1=xyz"))
-            .header("Cookie", "name=value")
-            .buildRequest(MockServletContext())
-        val sentryOptions = SentryOptions()
-        sentryOptions.isSendDefaultPii = false
-        val eventProcessor = fixture.getSut(request, sentryOptions)
-        val event = SentryEvent()
-
-        eventProcessor.process(event, null)
-
-        assertNotNull(event.request) {
-            assertNull(it.cookies)
-        }
-    }
-
-    @Test
-    fun `when sendDefaultPii is set to false, does not attach sensitive headers`() {
-        val request = MockMvcRequestBuilders
-            .get(URI.create("http://example.com?param1=xyz"))
-            .header("some-header", "some-header value")
-            .header("X-FORWARDED-FOR", "192.168.0.1")
-            .header("authorization", "Token")
-            .header("Authorization", "Token")
-            .header("Cookie", "some cookies")
-            .buildRequest(MockServletContext())
-        val sentryOptions = SentryOptions()
-        sentryOptions.isSendDefaultPii = false
-        val eventProcessor = fixture.getSut(request, sentryOptions)
-        val event = SentryEvent()
-
-        eventProcessor.process(event, null)
-
-        assertNotNull(event.request) {
-            assertFalse(it.headers.containsKey("X-FORWARDED-FOR"))
-            assertFalse(it.headers.containsKey("Authorization"))
-            assertFalse(it.headers.containsKey("authorization"))
-            assertFalse(it.headers.containsKey("Cookie"))
-            assertTrue(it.headers.containsKey("some-header"))
-        }
-    }
 
     @Test
     fun `when event does not have transaction name, sets the transaction name from the current request`() {
